+++ workspace/projects/demo/src/app/app.component.html
@@ -1,2 +1,27 @@
-<app-template-driven-form></app-template-driven-form>
-<app-reactive-form></app-reactive-form>
+<h1>DateValueAccessor Demo</h1>
+
+<p>
+  This project knows a number of different value accessors.
+</p>
+
+<ol>
+  <li>If you want to use date objects, select the first option.</li>
+  <li>If you want to work with ISO-formatted strings instead, select the second option.</li>
+</ol>
+
+<details class="clearfix">
+  <summary><b>1. Examples for working with date objects</b> (click to show)</summary>
+
+  <app-template-driven-form></app-template-driven-form>
+  <app-reactive-form></app-reactive-form>
+
+</details>
+
+
+<details>
+  <summary><b>2. Examples for working with ISO-formatted strings</b> (click to show)</summary>
+
+  <app-template-driven-form></app-template-driven-form>
+  <app-reactive-form></app-reactive-form>
+
+</details>
