+++ workspace/projects/demo/src/app/demo-with-dates/reactive-form/reactive-form.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
-import { Release } from '../shared/release';
+import { Release } from '../../shared/release';
 
 @Component({
   selector: 'app-reactive-form',
